Display each player's cards
Add `while` and `for` loop to contain the main code for the game logic.
Add `is_any_player_in` function for testing whether at least one player is currently in the game.
Moved code for printing list of cards from `__str__` method of Deck class into a new top-level function, so the code can be used with players' lists of cards as well.